myra's update to comments
